changed to uintptr_t and consistent macro use

+++ cpp/daal/src/services/service_topo.h
@@ -61,9 +61,9 @@ typedef cpuset_t cpu_set_t;
         #ifdef __x86_64__
             #define LNX_PTR2INT unsigned long long
             #define LNX_MY1CON  1LL
-        #elif __aarch64__
-using LNX_PTR2INT                = unsigned long long;
-constexpr LNX_PTR2INT LNX_MY1CON = 1LL;
+        #elif __ARM_ARCH
+            using LNX_PTR2INT                = uintptr_t;
+            constexpr LNX_PTR2INT LNX_MY1CON = 1LL;
         #else
             #define LNX_PTR2INT unsigned int
             #define LNX_MY1CON  1
